+++ src/api/index.ts
@@ -1,11 +1,12 @@
 import axios from 'axios';
 
+import { OWHAT_TOKEN } from '~/constant/token';
 import { BrowserStorage } from '~/utils/storage';
 
 const BASE_URL = import.meta.env.VITE_BASE_URL;
 const PORT = import.meta.env.VITE_PORT;
 
-const myStorage = new BrowserStorage<string>('OWHAT_TOKEN');
+const myStorage = new BrowserStorage<string>(OWHAT_TOKEN);
 const token = myStorage.get();
 
 export const instance = axios.create({
