+++ src/constant/token.ts
@@ -1,0 +1,1 @@
+export const OWHAT_TOKEN = 'OWHAT_TOKEN';
